:bug: Crash en el Widget Mediano, crash al actualizar el horario, bug en el que si el SAES agregaba 2 veces la misma clase en el mismo dia a la misma hora causaba un error y no mostraba el horario

+++ app/src/main/java/ziox/ramiro/saes/features/saes/features/schedule/view_models/ScheduleViewModel.kt
@@ -46,11 +46,13 @@ class ScheduleViewModel(
             it.id == classSchedule.id
         }
 
-        runOnDefaultThread {
-            customScheduleRoomRepository.removeClass(classSchedule.id)
-            customScheduleRoomRepository.addClass(classSchedule)
-        }
+        if(index in scheduleList.indices){
+            runOnDefaultThread {
+                customScheduleRoomRepository.removeClass(classSchedule.id)
+                customScheduleRoomRepository.addClass(classSchedule)
+            }
 
-        scheduleList[index] = classSchedule.toClassSchedule()
+            scheduleList[index] = classSchedule.toClassSchedule()
+        }
     }
 }
